refactor(user): rename user/DTOs dir to user/dtos
