+++ src/modules/user/useCases/registerUser/RegisterUserUseCase.ts
@@ -5,7 +5,7 @@ import { v4 } from 'uuid';
 import { IHashProvider } from '@modules/user/providers/HashProvider/models/IHashProvider';
 import { User } from '../../models/User';
 import { IUserRepository } from '../../repositories/IUserRepository';
-import { RegisterUserDTO } from '../../DTOs/RegisterUserDTO';
+import { RegisterUserDTO } from '../../dtos/RegisterUserDTO';
 
 @injectable()
 export class RegisterUserUseCase {
